Create 6-square.py and 6-main.py
Create class Square that defines a square  by private instance attributes size and position, with their corresponding  getter and type-checked setter for both. It offers also public instance methods area (to calculate and return the area of the square) and my_print to print the square os size size at the position position.